New script to analyze perfcounters collected with interception library

+++ tracers/hcTracer.h
@@ -18,7 +18,7 @@ TRACEPOINT_EVENT(
         const char*, name_arg,
         const char*, long_name_arg,
         const char*, tf_name_arg,
-        uint32_t*, workgroup_size_arg,
+        uint16_t*, workgroup_size_arg,
         uint32_t*, grid_size_arg,
         uint64_t, static_group_segment_size_arg,
         uint64_t, private_segment_size_arg,
@@ -30,7 +30,7 @@ TRACEPOINT_EVENT(
         ctf_string(name, name_arg)
         ctf_string(long_name, long_name_arg)
         ctf_string(tf_name, tf_name_arg)
-        ctf_array(uint32_t, workgroup_size, workgroup_size_arg, 3)
+        ctf_array(uint16_t, workgroup_size, workgroup_size_arg, 3)
         ctf_array(uint32_t, grid_size, grid_size_arg, 3)
         ctf_integer(uint64_t, static_group_segment_size, static_group_segment_size_arg)
         ctf_integer(uint64_t, private_segment_size, private_segment_size_arg)
@@ -46,7 +46,7 @@ TRACEPOINT_EVENT(
         const char*, name_arg,
         const char*, long_name_arg,
         const char*, tf_name_arg,
-        uint32_t*, workgroup_size_arg,
+        uint16_t*, workgroup_size_arg,
         uint32_t*, grid_size_arg,
         uint64_t, static_group_segment_size_arg,
         uint64_t, private_segment_size_arg,
@@ -58,7 +58,7 @@ TRACEPOINT_EVENT(
         ctf_string(name, name_arg)
         ctf_string(long_name, long_name_arg)
         ctf_string(tf_name, tf_name_arg)
-        ctf_array(uint32_t, workgroup_size, workgroup_size_arg, 3)
+        ctf_array(uint16_t, workgroup_size, workgroup_size_arg, 3)
         ctf_array(uint32_t, grid_size, grid_size_arg, 3)
         ctf_integer(uint64_t, static_group_segment_size, static_group_segment_size_arg)
         ctf_integer(uint64_t, private_segment_size, private_segment_size_arg)
